chore(internal): ensure all frames reported when wrapping (#5081)
This change adds the new `RESUME` instruction, introduced in Python
3.11, to the wrapping bytecode to correctly represent the current call
stack, in line with previous implementations. The new opcode with
argument `0` is used to indicate the beginning of a function. Without
it, the call stack is missing the entry corresponding to the new
bytecode object that is being swapped in for the original one, thus
making predictions about the frames at the top of the stack inconsistent
across CPython versions.