fix(gevent): unload time but not typing modules for module cloning (#5135)
This PR makes a fix to #4863 which inadvertently removed unloading the
`time` module as well as removing the `typing` module from the list of
standard library modules to not unload for module cloning. We found that
the `time` module is a special case (which is important to unload due to
interactions with gevent patching) as it is implicitly imported by
recent versions of CPython on interpreter startup, so it may be present
in `sys.modules` already.
The `typing` module is also problematic on being reloaded for older
versions of Python < 3.7, so we should not unload/reload this module as
well.