fix(aws_lambda): update signal handler to avoid wrapping multiple times (#5190)
## Description
Fix a bug in which our signal might be wrapped multiple times, which
cause it to call the the signal multiple times.

## Testing strategy
Updated unit tests which were causing this, I first added them in #5176,
moved them here.

## Risk
This was only affecting unit testing, since this wouldn't have happened
in production. This is because in testing we are setting the same
handler for multiple tests, which the system would recognize as
previously defined. In an AWS Lambda function, once the environment is
dead, all signals are lost. So there will not be multiple wrapping.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Author is aware of the performance implications of this PR as
reported in the benchmarks PR comment.

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer is aware of, and discussed the performance implications
of this PR as reported in the benchmarks PR comment.