fix(elasticsearch): omit large body tag values (#5229)
The elasticsearch body tag value in practice can be very large. Since
the value is not truncated, it can easily result in traces exceeding the
max size limit which results in dropped traces.

The proposed solution is to omit the value. This is because obfuscation
logic lives in the Agent and so the body cannot be truncated as
sensitive data would be leaked. No body tag is better than no trace.

Ideally the obfuscation logic should exist in the client so that the
body can be truncated.