fix(debugging): handle disabled probes via RCM (#5282)
This change fixes the handling of disabled configuration targets in
dynamic instrumentation. Because of the currently broken RCM callback
signature, the disable events were not handled correctly and caused an
exception to be thrown during processing of the RCM event. We fix this
by working around the problematic callback signature.

Testing is performed via the backend demo job monitoring, which is what
allowed the issue to surface in the first place.