feat(celery): trace celery's "beat" scheduling service (#5327)
This pull request adds to the Celery integration support for
`celery.beat`. This is a module that can trigger Celery tasks on a
schedule. Supporting it in the integration means that users will have
more visibility into the conditions under which their tasks are
triggered and will get traces when Beat scheduling fails.

Co-authored-by: Yun Kim <35776586+Yun-Kim@users.noreply.github.com>