fix(tracing): honor sample rate set by agent (#5339)
This change fixes a bug causing sample rate data from the agent not to
be honored in the tracer's sampling behavior. The bug is that the
sampling decision is made before the `env` tag is set. This tag is a key
piece of information necessary to make sampling decisions based on
responses from the agent.

The change also includes a regression test.