chore(pytest): move CI visibility initialization to start of pytest command (#5466)
This PR moves the CI Visibility enabling step to the `pytest_configure`
hook instead of the `pytest_sessionstart` hook so that we can enable CI
Visibility as early as possible in the pytest run command. See
https://github.com/DataDog/dd-trace-py/pull/5446/files#r1156967374 for
more context.

This is because the planned test suite level visibility for CI
Visibility aims to track the entire pytest command (including test
command, test modules, test functions) instead of our current test
function traces. This requires the ddtrace pytest plugin to enable the
CIRecorder at the earliest possible step in the pytest test command.

This functionally does not change the current behavior of the pytest
plugin, since the `pytest_configure` and `pytest_sessionstart` hooks
happen almost immediately after each other in the pytest command
lifecycle. All this does is to lay the groundwork for test suite level
visibility work for CI Visibility. **Note that this change does not
implement the test suite level visibility for CI Visibility, but is a
pre-requisite for it.**

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] PR description includes explicit acknowledgement/acceptance of the
performance implications of this PR as reported in the benchmarks PR
comment.

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.

---------

Co-authored-by: Zachary Groves <32471391+ZStriker19@users.noreply.github.com>