chore(tracing): encapsulate agent-specific writer functionality (#5475)
This change facilitates the creation of new `*Writer` classes by
encapsulating the behavior of the `AgentWriter` that is specific to
communication with the Datadog Agent in its own class. This leaves most
of the writer's functionality, which could apply to any trace writing
workload that uses HTTP, in the new `HTTPWriter` class. This change
builds on the approach from
https://github.com/DataDog/dd-trace-py/pull/3690.

https://github.com/DataDog/dd-trace-py/pull/5471 builds on this change
to implement a new `CIVisibilityWriter` class.

This change only affects code organization, not functionality, so
existing tests are sufficient to cover the risk.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] PR description includes explicit acknowledgement/acceptance of the
performance implications of this PR as reported in the benchmarks PR
comment.

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.