chore(ci): add special branches for testing build_deploy, fix lib-inject job (#5486)
In order to test build_deploy jobs a special branch pattern is added.

Github doesn't verify that jobs adhere to its workflow schema until the
job matches the `on` criteria and is run. Previously this workflow only
ran on release branches and on releases. This meant that there was no
way to ensure the jobs actually parsed correctly before a PR was merged
or a release was done.


To demonstrate the fix, a bug exists in the lib-injection job which
results in


![image](https://user-images.githubusercontent.com/6321485/230226834-4e5bc217-9dd9-4925-8d0c-cf75c4657910.png)

on the first commit of this PR.

A commit is pushed with the fix which shows that the workflow at least
parses:


![image](https://user-images.githubusercontent.com/6321485/230227458-1c2ef63c-48f0-4ee7-8795-780ffc68eb4a.png)


This of course doesn't verify that the job runs successfully but it at
least lets us know that it loads fine (which given Github's workflow
rules and lack of documentation goes a long way).