fix(ci-visibility): add dd_origin propagation to CIVisibilityEncoder (#5541)
This PR builds on #5489 by adding dd_origin propagation to the
CIVisibilityEncoder, which is necessary for CI Visibility. This is not a
new feature as this functionality is already present in the regular
encoder.

This doesn't need a changelog since this is internal and is a fix to an
unreleased change.